+++ mmyolo/models/detectors/yolo_detector.py
@@ -1,5 +1,4 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from typing import Union
 
 import torch
 from mmdet.models.detectors.single_stage import SingleStageDetector
@@ -53,20 +52,3 @@ class YOLODetector(SingleStageDetector):
         if use_syncbn and get_world_size() > 1:
             torch.nn.SyncBatchNorm.convert_sync_batchnorm(self)
             print_log('Using SyncBatchNorm()', 'current')
-
-    def assign(self, data: dict) -> Union[dict, list]:
-        """Calculate assigning results from a batch of inputs and data
-        samples.This function is provided to the `show_assign.py` script.
-
-        Args:
-            data (dict or tuple or list): Data sampled from dataset.
-
-        Returns:
-            dict: A dictionary of assigning components.
-        """
-        assert isinstance(data, dict)
-        assert len(data['inputs']) == 1, 'Only support batchsize == 1'
-        data = self.data_preprocessor(data, True)
-        inputs_hw = data['inputs'].shape[-2:]
-        assign_results = self.bbox_head.assign(data['data_samples'], inputs_hw)
-        return assign_results
