+++ projects/showassignment/dense_heads/__init__.py
@@ -1,3 +1,3 @@
-from .yolov5_head_assigner import YOLOv5HeadAssigner
+from .yolov5_head_assigner import YOLOv5HeadShowAssigner
 
-__all__ = ['YOLOv5HeadAssigner']
+__all__ = ['YOLOv5HeadShowAssigner']
