+++ projects/showassignment/detectors/__init__.py
@@ -1,0 +1,4 @@
+from projects.showassignment.detectors.yolo_detector_showassignment import \
+    YOLODetectorShowAssigner
+
+__all__ = ['YOLODetectorShowAssigner']
